Updated demand data generation
Lots of updates, but its v1 so its cool
